requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-ldap==3.4.4
python-jose[cryptography]==3.3.0
asyncpg==0.29.0
httpx==0.27.0
pydantic==2.7.1
python-dotenv==1.0.0
